// src/app/models/actions.ts
export class Actions {
    PopularMovies:string = "POPULAR/MOVIES";
    PopularSeries:string = "POPULAR/SERIES";
    SearchMovie:string = "SEARCH/MOVIE";
    SearchSerie:string = "SEARCH/SERIE";
    DetailMovie: string = "DETAIL/MOVIE";
    DetailSerie: string = "DETAIL/SERIE";
    MovieCredits: string = "CREDITS/MOVIE";
    SerieCredits: string = "CREDITS/SERIE";
    Credit: string = "CREDIT";
    Person: string = "PERSON/PEOPLE";
    People: string = "POPULAR/PEOPLE";
    SearchPeople: string = "SEARCH/PEOPLE";
    TrendingAll: string = "TRENDING/ALL";
    TrendingMovie: string = "TRENDING/MOVIE";
    TrendingSerie: string = "TRENDING/SERIE";
    TrendingPeople: string = "TRENDING/PEOPLE";
}